+++ src/dispatcher/index.spec.js
@@ -1,4 +1,3 @@
-process.env.RABBITHOST = '46.4.8.29';
 import {Dispatcher} from './dispatcher';
 
 test('if dispatcher class exists', () => {
