+++ src/download-file/downloader.js
@@ -2,6 +2,8 @@ import http from 'http';
 import https from 'https';
 import fs from 'fs';
 import logger from './child-logger.js';
+import path from 'path';
+import { URL } from 'url';
 
 /**
  * Downloads a file from the internet to the local filesystem.
@@ -54,4 +56,41 @@ const downloadFile = async (url, downloadPath, options = {}) => {
     })
 }
 
-export default downloadFile;
+/**
+ * Downloads data into the given target from the given URL.
+ * Modifies the given job object in place with status and results.
+ * @param {Object} workerJob The job object
+ * @param {Array} inputs The inputs for this process
+ *   First input is the full URL to download from
+ *   Second input is the absolute path, including filename and suffic, to store the download to
+ *   (optional) Third input is the username for basic auth
+ *   (optional) Fourth input is the password for basic auth
+ */
+const callbackWorker = async (workerJob, inputs) => {
+    const uri = inputs[0];
+    const downloadPath = inputs[1];
+    const username = inputs[2];
+    const password = inputs[3];
+    const url = new URL(uri);
+
+    logger.info(`Downloading ${url.href} ...`);
+
+    // if provided: add basic auth credentials to request option
+    const options = {};
+    if (username && password) {
+        logger.debug('Using credentials to download file.')
+        options.auth = `${username}:${password}`
+    }
+
+    const downloadDir = path.dirname(downloadPath);
+    await fs.mkdirSync(downloadDir, { recursive: true });
+
+    return downloadFile(url, downloadPath, options)
+        .then((downloadPath) => {
+            logger.info('Download has finished.');
+            workerJob.status = 'success';
+            workerJob.outputs = [downloadPath];
+        });
+};
+
+export { downloadFile, callbackWorker };
