--- src/download-file/index.spec.js
@@ -1,18 +1,0 @@
-process.env.RABBITHOST = '46.4.8.29';
-import callbackWorker from './index';
-
-test('if function exists', () => {
-    expect(callbackWorker).toBeDefined();
-});
-
-test('if worker provides correct outputs', async () => {
-    const job = {};
-    const outputPath = '/tmp/README.md';
-    await callbackWorker(job, [
-        'https://raw.githubusercontent.com/klips-project/rabbitmq-worker/main/README.md',
-        outputPath
-    ]);
-    expect(job.outputs).toBeDefined();
-    expect(job.outputs.length).toBe(1);
-    expect(job.outputs[0]).toBe(outputPath);
-});
