+++ src/geoserver-publish-geotiff/index.spec.js
@@ -1,4 +1,3 @@
-process.env.RABBITHOST = '46.4.8.29';
 process.env.GEOSERVER_REST_URL = 'http://46.4.8.29:8080/geoserver/';
 import index from './index';
